optimizer: rename optimizeex to hoptimize
