+++ pkg/cmd/hoptimize.go
@@ -4,29 +4,27 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/c9s/bbgo/pkg/data/tsv"
 	"github.com/c9s/bbgo/pkg/optimizer"
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3"
-	"io"
 	"io/ioutil"
 	"os"
 	"time"
 )
 
 func init() {
-	optimizeExCmd.Flags().String("optimizer-config", "optimizer.yaml", "config file")
-	optimizeExCmd.Flags().String("name", "", "assign a name to the study")
-	optimizeExCmd.Flags().Bool("json-keep-all", false, "keep all results of trials")
-	optimizeExCmd.Flags().String("output", "output", "backtest report output directory")
-	optimizeExCmd.Flags().Bool("json", false, "print optimizer metrics in json format")
-	optimizeExCmd.Flags().Bool("tsv", false, "print optimizer metrics in csv format")
-	RootCmd.AddCommand(optimizeExCmd)
+	hoptimizeCmd.Flags().String("optimizer-config", "optimizer.yaml", "config file")
+	hoptimizeCmd.Flags().String("name", "", "assign an optimization session name")
+	hoptimizeCmd.Flags().Bool("json-keep-all", false, "keep all results of trials")
+	hoptimizeCmd.Flags().String("output", "output", "backtest report output directory")
+	hoptimizeCmd.Flags().Bool("json", false, "print optimizer metrics in json format")
+	hoptimizeCmd.Flags().Bool("tsv", false, "print optimizer metrics in csv format")
+	RootCmd.AddCommand(hoptimizeCmd)
 }
 
-var optimizeExCmd = &cobra.Command{
-	Use:   "optimizeex",
+var hoptimizeCmd = &cobra.Command{
+	Use:   "hoptimize",
 	Short: "run hyperparameter optimizer (experimental)",
 
 	// SilenceUsage is an option to silence usage when an error occurs.
@@ -43,7 +41,7 @@ var optimizeExCmd = &cobra.Command{
 			return err
 		}
 
-		studyName, err := cmd.Flags().GetString("name")
+		optSessionName, err := cmd.Flags().GetString("name")
 		if err != nil {
 			return err
 		}
@@ -94,10 +92,10 @@ var optimizeExCmd = &cobra.Command{
 		defer cancel()
 		_ = ctx
 
-		if len(studyName) == 0 {
-			studyName = fmt.Sprintf("bbgo-hpopt-%v", time.Now().UnixMilli())
+		if len(optSessionName) == 0 {
+			optSessionName = fmt.Sprintf("bbgo-hpopt-%v", time.Now().UnixMilli())
 		}
-		tempDirNameFormat := fmt.Sprintf("%s-config-*", studyName)
+		tempDirNameFormat := fmt.Sprintf("%s-config-*", optSessionName)
 		configDir, err := os.MkdirTemp("", tempDirNameFormat)
 		if err != nil {
 			return err
@@ -112,8 +110,8 @@ var optimizeExCmd = &cobra.Command{
 		}
 
 		optz := &optimizer.HyperparameterOptimizer{
-			StudyName: studyName,
-			Config:    optConfig,
+			SessionName: optSessionName,
+			Config:      optConfig,
 		}
 
 		if err := executor.Prepare(configJson); err != nil {
@@ -137,13 +135,13 @@ var optimizeExCmd = &cobra.Command{
 			// print report JSON to stdout
 			fmt.Println(string(out))
 		} else if printTsvFormat {
-			if err := formatResultsTsv(os.Stdout, report.Parameters, report.Trials); err != nil {
+			if err := optimizer.FormatResultsTsv(os.Stdout, report.Parameters, report.Trials); err != nil {
 				return err
 			}
 		} else {
 			color.Green("OPTIMIZER REPORT")
 			color.Green("===============================================\n")
-			color.Green("STUDY NAME: %s\n", report.Name)
+			color.Green("SESSION NAME: %s\n", report.Name)
 			color.Green("OPTIMIZE OBJECTIVE: %s\n", report.Objective)
 			color.Green("BEST OBJECTIVE VALUE: %s\n", report.Best.Value)
 			color.Green("OPTIMAL PARAMETERS:")
@@ -159,45 +157,4 @@ var optimizeExCmd = &cobra.Command{
 
 		return nil
 	},
-}
-
-func formatResultsTsv(writer io.WriteCloser, labelPaths map[string]string, results []*optimizer.HyperparameterOptimizeTrialResult) error {
-	headerLen := len(labelPaths)
-	headers := make([]string, 0, headerLen)
-	for label := range labelPaths {
-		headers = append(headers, label)
-	}
-
-	rows := make([][]interface{}, len(labelPaths))
-	for ri, result := range results {
-		row := make([]interface{}, headerLen)
-		for ci, columnKey := range headers {
-			var ok bool
-			if row[ci], ok = result.Parameters[columnKey]; !ok {
-				return fmt.Errorf(`missing parameter "%s" from trial result (%v)`, columnKey, result.Parameters)
-			}
-		}
-		rows[ri] = row
-	}
-
-	w := tsv.NewWriter(writer)
-	if err := w.Write(headers); err != nil {
-		return err
-	}
-
-	for _, row := range rows {
-		var cells []string
-		for _, o := range row {
-			cell, err := castCellValue(o)
-			if err != nil {
-				return err
-			}
-			cells = append(cells, cell)
-		}
-
-		if err := w.Write(cells); err != nil {
-			return err
-		}
-	}
-	return w.Close()
 }
