+++ pkg/cmd/optimize.go
@@ -4,16 +4,12 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
-	"strconv"
 
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3"
 
-	"github.com/c9s/bbgo/pkg/data/tsv"
-	"github.com/c9s/bbgo/pkg/fixedpoint"
 	"github.com/c9s/bbgo/pkg/optimizer"
 )
 
@@ -119,7 +115,7 @@ var optimizeCmd = &cobra.Command{
 			// print metrics JSON to stdout
 			fmt.Println(string(out))
 		} else if printTsvFormat {
-			if err := formatMetricsTsv(metrics, os.Stdout); err != nil {
+			if err := optimizer.FormatMetricsTsv(os.Stdout, metrics); err != nil {
 				return err
 			}
 		} else {
@@ -137,96 +133,4 @@ var optimizeCmd = &cobra.Command{
 
 		return nil
 	},
-}
-
-func transformMetricsToRows(metrics map[string][]optimizer.Metric) (headers []string, rows [][]interface{}) {
-	var metricsKeys []string
-	for k := range metrics {
-		metricsKeys = append(metricsKeys, k)
-	}
-
-	var numEntries int
-	var paramLabels []string
-	for _, ms := range metrics {
-		for _, m := range ms {
-			paramLabels = m.Labels
-			break
-		}
-
-		numEntries = len(ms)
-		break
-	}
-
-	headers = append(paramLabels, metricsKeys...)
-	rows = make([][]interface{}, numEntries)
-
-	var metricsRows = make([][]interface{}, numEntries)
-
-	// build params into the rows
-	for i, m := range metrics[metricsKeys[0]] {
-		rows[i] = m.Params
-	}
-
-	for _, metricKey := range metricsKeys {
-		for i, ms := range metrics[metricKey] {
-			if len(metricsRows[i]) == 0 {
-				metricsRows[i] = make([]interface{}, 0, len(metricsKeys))
-			}
-			metricsRows[i] = append(metricsRows[i], ms.Value)
-		}
-	}
-
-	// merge rows
-	for i := range rows {
-		rows[i] = append(rows[i], metricsRows[i]...)
-	}
-
-	return headers, rows
-}
-
-func formatMetricsTsv(metrics map[string][]optimizer.Metric, writer io.WriteCloser) error {
-	headers, rows := transformMetricsToRows(metrics)
-	w := tsv.NewWriter(writer)
-	if err := w.Write(headers); err != nil {
-		return err
-	}
-
-	for _, row := range rows {
-		var cells []string
-		for _, o := range row {
-			cell, err := castCellValue(o)
-			if err != nil {
-				return err
-			}
-			cells = append(cells, cell)
-		}
-
-		if err := w.Write(cells); err != nil {
-			return err
-		}
-	}
-	return w.Close()
-}
-
-func castCellValue(a interface{}) (string, error) {
-	switch tv := a.(type) {
-	case fixedpoint.Value:
-		return tv.String(), nil
-	case float64:
-		return strconv.FormatFloat(tv, 'f', -1, 64), nil
-	case int64:
-		return strconv.FormatInt(tv, 10), nil
-	case int32:
-		return strconv.FormatInt(int64(tv), 10), nil
-	case int:
-		return strconv.Itoa(tv), nil
-	case bool:
-		return strconv.FormatBool(tv), nil
-	case string:
-		return tv, nil
-	case []byte:
-		return string(tv), nil
-	default:
-		return "", fmt.Errorf("unsupported object type: %T value: %v", tv, tv)
-	}
 }
